+++ app/Views/pendaftaran_haji_khusus/laporan_jamaah.php
@@ -48,7 +48,7 @@
                                           <td><?=$row->jamaah_ttl;?></td>
                                           <td><?=$row->jamaah_usia;?></td>
                                           <td>
-                                            <a style="width:15px;" class="btn btn-danger" href="<?php echo base_url();?>/pendaftaran_haji_khusus/laporan/hapus/<?=$row->jamaah_id;?>"><i class="fa fa-times"></i></a>
+                                            <a style="width:15px;" class="btn btn-danger" href="<?php echo base_url();?>/pendaftaran_haji_khusus/laporan/hapus/<?=$row->jamaah_id;?>" onclick="javascript: return confirm('Yakin Menghapus Data?')"><i class="fa fa-times"></i></a>
                                           </td>
                                       </tr>
                                       <?php
